// pages/api/reviews/review-interface.ts
export default interface IReview {
  slug: string;
  name: string;
  aboutProduct: string;
  advantage: string;
  disadvantages: string;
  nameUser: string;
  userCity: string;
  impression: string;
  reliability: string;
  functionality: string;
  quality: string;
  photoMatching: string;
  recommend: boolean;
  discommend: boolean;
  periodOfUseUser: string;
  frequencyOfUseUser: string;
  policyData: boolean;
  timestamps: boolean;
}